chore: separate analytics from building model
Refs: #19
