+++ src/frontend/Dockerfile
@@ -6,8 +6,8 @@ WORKDIR /app
 # Copy package files first for better caching
 COPY package*.json ./
 
-# Install dependencies (this layer will be cached if package*.json don't change)
-RUN npm ci --silent && \
+# Install dependencies
+RUN npm install && \
     npm cache clean --force
 
 # Copy source code
